Move CanvasTreeView to separate file.
