+++ src/apps/LayoutBuilder/Preview/views/index.ts
@@ -1,3 +1,4 @@
 export { CssView } from './CssView';
 export { HtmlView } from './HtmlView';
 export { TreeView } from './TreeView';
+export { CanvasTreeView } from './CanvasTreeView';
